+++ src/main/java/Duke/Command.java
@@ -1,4 +1,6 @@
-// Command interface
+package Duke;
+
+// Commands.Command interface
 public interface Command {
     void execute(TaskList tasks, Ui ui, Storage storage) throws DukeException;
 }
