+++ src/main/java/Duke/DeleteCommand.java
@@ -1,9 +1,11 @@
+package Duke;
+
 import java.io.IOException;
 
-public class UnmarkCommand implements Command {
+public class DeleteCommand implements Command {
     private int taskIndex;
 
-    public UnmarkCommand(int taskIndex) {
+    public DeleteCommand(int taskIndex) {
         this.taskIndex = taskIndex;
     }
 
@@ -11,10 +13,10 @@ public class UnmarkCommand implements Command {
     public void execute(TaskList tasks, Ui ui, Storage storage) {
         try {
             if (taskIndex >= 0 && taskIndex < (tasks.getTasks()).size()) {
-                Task markedTask = tasks.getTask(taskIndex);
-                tasks.markTaskAsNotDone(taskIndex);
+                Task taskBeforeDeletion = tasks.getTask(taskIndex);
+                tasks.deleteTask(taskIndex);
                 storage.saveTasksToFile(tasks);
-                ui.showUnmarkTaskAsDone(markedTask);
+                ui.showDeleteMessage(tasks.getTasks(), taskBeforeDeletion);
             } else {
                 ui.invalidTaskIndex();
             }
@@ -23,3 +25,4 @@ public class UnmarkCommand implements Command {
         }
     }
 }
+
