+++ src/main/java/Duke/Duke.java
@@ -1,3 +1,5 @@
+package Duke;
+
 import java.io.IOException;
 import java.io.FileNotFoundException;
