+++ src/main/java/Duke/ExitCommand.java
@@ -1,4 +1,6 @@
-// ExitCommand class
+package Duke;
+
+// Commands.ExitCommand class
 public class ExitCommand implements Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
