+++ src/main/java/Duke/FindCommand.java
@@ -1,3 +1,5 @@
+package Duke;
+
 public class FindCommand implements Command {
     private String keyword;
     public FindCommand(String keyword) {
