+++ src/main/java/Duke/MarkCommand.java
@@ -1,3 +1,5 @@
+package Duke;
+
 import java.io.IOException;
 
 public class MarkCommand implements Command {
