+++ src/main/java/Duke/Parser.java
@@ -1,4 +1,5 @@
-import java.io.IOException;
+package Duke;
+
 import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
