+++ src/main/java/Duke/Storage.java
@@ -1,13 +1,11 @@
+package Duke;
+
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.nio.file.*;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
-import java.util.List;
-
-import java.nio.file.*;
-import java.io.IOException;
 import java.util.List;
 
 public class Storage {
